# storage: add prometheus metrics for seeders/leechers