@mrd0ll4r 2017-05-03
Some new benchmarks:

```
benchmark                               old ns/op     new ns/op     delta
BenchmarkPut                            387           398           +2.84%
BenchmarkPut1k                          403           458           +13.65%
BenchmarkPut1kInfohash                  461           482           +4.56%
BenchmarkPut1kInfohash1k                466           481           +3.22%
BenchmarkPutDelete                      1302          1326          +1.84%
BenchmarkPutDelete1k                    1293          1413          +9.28%
BenchmarkPutDelete1kInfohash            1325          1407          +6.19%
BenchmarkPutDelete1kInfohash1k          1326          1378          +3.92%
BenchmarkDeleteNonexist                 199           208           +4.52%
BenchmarkDeleteNonexist1k               205           209           +1.95%
BenchmarkDeleteNonexist1kInfohash       209           218           +4.31%
BenchmarkDeleteNonexist1kInfohash1k     213           219           +2.82%
BenchmarkPutGradDelete                  1866          2166          +16.08%
BenchmarkPutGradDelete1k                1863          2149          +15.35%
BenchmarkPutGradDelete1kInfohash        1960          2235          +14.03%
BenchmarkPutGradDelete1kInfohash1k      2010          2263          +12.59%
BenchmarkGradNonexist                   394           453           +14.97%
BenchmarkGradNonexist1k                 416           502           +20.67%
BenchmarkGradNonexist1kInfohash         468           502           +7.26%
BenchmarkGradNonexist1kInfohash1k       467           514           +10.06%
BenchmarkAnnounceLeecher                17014         17395         +2.24%
BenchmarkAnnounceLeecher1kInfohash      20581         20881         +1.46%
BenchmarkAnnounceSeeder                 17028         18069         +6.11%
BenchmarkAnnounceSeeder1kInfohash       21082         21186         +0.49%
```
(Taken on different machine)

Not sure what's happening with GraduateLeecher, will have to investigate.